+++ src/main/java/ex_grupo11/ex_api_grupo11/responses/LaminaResponse.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class LaminaResponse {
-    private Integer idLamina;
+    private Long idLamina;
     private String nombreLamina;
     private TipoCartaResponse tipoCarta;
     private String imagen;
