+++ src/main/java/ex_grupo11/ex_api_grupo11/responses/TipoCartaResponse.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class TipoCartaResponse {
-    private Integer idTipo;
+    private Long idTipo;
     private String nombreTipo;
 }
